Fix/reconnector

See merge request ProtonVPN/linux/protonvpn-nm-core!88

+++ protonvpn_nm_lib/services/killswitch_manager.py
@@ -50,7 +50,7 @@ class KillSwitchManager(AbstractInterfaceManager):
             }
         }
         self.disable_connectivity_check()
-        self.update_connection_status()
+        logger.info("Initialized killswitch manager")
 
     def manage(self, action, is_menu=False, server_ip=None):
         """Manage killswitch.
@@ -69,6 +69,7 @@ class KillSwitchManager(AbstractInterfaceManager):
             )
         )
         self.disable_connectivity_check()
+        self.update_connection_status()
 
         if is_menu:
             if int(action) == KillswitchStatusEnum.HARD:
@@ -84,10 +85,20 @@ class KillSwitchManager(AbstractInterfaceManager):
 
             return
 
-        if action == "pre_connection":
+        if (
+            action == "pre_connection"
+        ) and (
+            self.interface_state_tracker[self.ks_conn_name]["is_running"]
+            and not self.interface_state_tracker[self.routed_conn_name]["exists"] # noqa
+        ):
             self.create_routed_connection(server_ip)
             self.deactivate_connection(self.ks_conn_name)
-        elif action == "post_connection":
+        elif (
+            action == "post_connection"
+        ) and (
+            not self.interface_state_tracker[self.ks_conn_name]["is_running"]
+            and self.interface_state_tracker[self.routed_conn_name]["is_running"] # noqa
+        ):
             self.activate_connection(self.ks_conn_name)
             self.delete_connection(self.routed_conn_name)
         elif action == "soft_connection":
@@ -95,10 +106,6 @@ class KillSwitchManager(AbstractInterfaceManager):
             self.manage("post_connection")
         elif action == "disable":
             self.delete_all_connections()
-        else:
-            raise exceptions.KillswitchError(
-                "Incorrect option for killswitch manager"
-            )
 
     def create_killswitch_connection(self):
         """Create killswitch connection/interface."""
@@ -120,10 +127,10 @@ class KillSwitchManager(AbstractInterfaceManager):
         self.create_connection(
             self.ks_conn_name,
             "Unable to activate {}".format(self.ks_conn_name),
-            subprocess_command
+            subprocess_command, exceptions.CreateBlockingKillswitchError
         )
 
-    def create_routed_connection(self, server_ip):
+    def create_routed_connection(self, server_ip, try_route_addrs=False):
         """Create routed connection/interface.
 
         Args:
@@ -138,13 +145,14 @@ class KillSwitchManager(AbstractInterfaceManager):
 
         route_data = [str(ipv4) for ipv4 in subnet_list]
         route_data_str = ",".join(route_data)
+
         subprocess_command = ""\
             "nmcli c a type dummy ifname {interface_name} "\
             "con-name {conn_name} ipv4.method manual "\
             "ipv4.addresses {ipv4_addrs} "\
             "ipv6.method manual ipv6.addresses {ipv6_addrs} "\
             "ipv6.gateway {ipv6_gateway} "\
-            "ipv4.route-metric 98 ipv6.route-metric 98 "\
+            "ipv4.route-metric 97 ipv6.route-metric 97 "\
             "ipv4.routes {routes}".format(
                 conn_name=self.routed_conn_name,
                 interface_name=self.routed_interface_name,
@@ -154,19 +162,43 @@ class KillSwitchManager(AbstractInterfaceManager):
                 routes=route_data_str
             ).split(" ")
 
-        self.create_connection(
-            self.routed_conn_name,
-            "Unable to activate {}".format(self.routed_conn_name),
-            subprocess_command
-        )
+        if try_route_addrs:
+            subprocess_command = ""\
+                "nmcli c a type dummy ifname {interface_name} "\
+                "con-name {conn_name} ipv4.method manual "\
+                "ipv4.addresses {ipv4_addrs} "\
+                "ipv6.method manual ipv6.addresses {ipv6_addrs} "\
+                "ipv6.gateway {ipv6_gateway} "\
+                "ipv4.route-metric 97 ipv6.route-metric 97".format(
+                    conn_name=self.routed_conn_name,
+                    interface_name=self.routed_interface_name,
+                    ipv4_addrs=route_data_str,
+                    ipv6_addrs=self.ipv6_dummy_addrs,
+                    ipv6_gateway=self.ipv6_dummy_gateway,
+                ).split(" ")
+
+        logger.info(subprocess_command)
+        exception_msg = "Unable to activate {}".format(self.routed_conn_name)
+
+        try:
+            self.create_connection(
+                self.routed_conn_name, exception_msg,
+                subprocess_command, exceptions.CreateRoutedKillswitchError
+            )
+        except exceptions.CreateRoutedKillswitchError as e:
+            if e.additional_context.returncode == 2 and not try_route_addrs:
+                return self.create_routed_connection(server_ip, True)
+            else:
+                raise exceptions.CreateRoutedKillswitchError(exception_msg)
 
     def create_connection(
-        self, conn_name, exception_msg, subprocess_command
+        self, conn_name, exception_msg,
+        subprocess_command, exception
     ):
         self.update_connection_status()
         if not self.interface_state_tracker[conn_name]["exists"]:
             self.run_subprocess(
-                exceptions.CreateKillswitchError,
+                exception,
                 exception_msg,
                 subprocess_command
             )
@@ -263,6 +295,8 @@ class KillSwitchManager(AbstractInterfaceManager):
             else:
                 self.interface_state_tracker[active_conn.get_id()]["is_running"] = True # noqa
 
+        logger.info("Tracker info: {}".format(self.interface_state_tracker))
+
     def run_subprocess(self, exception, exception_msg, *args):
         """Run provided input via subprocess.
 
@@ -291,7 +325,8 @@ class KillSwitchManager(AbstractInterfaceManager):
                 )
             )
             raise exception(
-                exception_msg
+                exception_msg,
+                subprocess_outpout
             )
 
     def disable_connectivity_check(self):
